+++ src/opt.c
@@ -250,11 +250,11 @@ static void next_addresses(block *address_block, block *input_block) {
     fill_block(zero2_block, address_block, address_block, 0);
 }
 
-uint64_t argon2i_index_size(const argon2_instance_t *instance) {
+uint32_t argon2i_index_size(const argon2_instance_t *instance) {
     return instance->segment_length * ARGON2_SYNC_POINTS - 2;
 }
 
-uint64_t argon2i_precompute(
+uint32_t argon2i_precompute(
     const argon2_instance_t *instance,
     argon2_precomputed_index_t *oIndex) {
